Remove file separation for now
